Add the missing timer device

+++ GlitchGame/Devices/Navigation.cs
@@ -26,11 +26,11 @@ namespace GlitchGame.Devices
             switch (machine.Registers[7])
             {
                 case 0: // get speed
-                    machine.Registers[7] = (short)(_body.LinearVelocity.X * Program.PixelsPerMeter);
-                    machine.Registers[8] = (short)(_body.LinearVelocity.Y * Program.PixelsPerMeter);
+                    machine.Registers[7] = (int)(_body.LinearVelocity.X * Program.PixelsPerMeter);
+                    machine.Registers[8] = (int)(_body.LinearVelocity.Y * Program.PixelsPerMeter);
                     break;
                 case 1: // get angular speed
-                    machine.Registers[7] = (short)(_body.AngularVelocity % Program.RadarRays);
+                    machine.Registers[7] = (int)(_body.AngularVelocity % Program.RadarRays);
                     break;
                 case 2: // get heading
                     machine.Registers[7] = Util.ToMachineRotation(_body.Rotation - ((float)Math.PI / 2));
